+++ app/pkg/libs/mrcalc/parallelepiped/format.go
@@ -25,12 +25,12 @@ func (f Format) String() string {
 }
 
 // IsValid - проверяет валиден ли текущий формат.
-func (f *Format) IsValid() bool {
+func (f Format) IsValid() bool {
 	return f.Length > 0 && f.Width > 0 && f.Height > 0
 }
 
 // BottomFormat - возвращает формат дна параллелепипеда.
-func (f *Format) BottomFormat() rect.Format {
+func (f Format) BottomFormat() rect.Format {
 	return rect.Format{
 		Width:  f.Length,
 		Height: f.Width,
@@ -38,11 +38,30 @@ func (f *Format) BottomFormat() rect.Format {
 }
 
 // IsZero - проверяет что формат является нулевым.
-func (f *Format) IsZero() bool {
+func (f Format) IsZero() bool {
 	return f.Length == 0 && f.Width == 0 && f.Height == 0
 }
 
 // Volume - возвращает объём формата.
-func (f *Format) Volume() float64 {
+func (f Format) Volume() float64 {
 	return f.Length * f.Width * f.Height
+}
+
+// Diff - возвращает разницу текущего формата и указанного.
+// Если какая-то сторона указанного формата больше той же стороны текущего формата,
+// то эта сторона будет приравнена 0.
+func (f Format) Diff(second Format) Format {
+	diff := func(first, second float64) float64 {
+		if first > second {
+			return first - second
+		}
+
+		return 0
+	}
+
+	return Format{
+		Length: diff(f.Length, second.Length),
+		Width:  diff(f.Width, second.Width),
+		Height: diff(f.Height, second.Height),
+	}
 }